+ PrepareSource.sh: Windows compil and Linux packages for MediaInfoLib

+++ MediaInfoLib/PrepareSource.sh
@@ -40,28 +40,39 @@ function _get_source () {
 function _linux_compil () {
 
     echo
-    echo "Create the archive for compilation on Linux:"
+    echo "Prepare the archive for compilation on Linux:"
     echo "1: copy what is wanted..."
 
     cd $Path
     mkdir MediaInfo_DLL_${Version}_GNU_FromSource
     cd MediaInfo_DLL_${Version}_GNU_FromSource
 
-    cp -r $ZL_source .
     cp -r $MIL_source .
     mv MediaInfoLib/Project/GNU/Library/AddThisToRoot_DLL_compile.sh SO_Compile.sh
-    # TODO: call the function who build a clean ZenLib compil archive
 
-    mkdir -p Shared/Project/_Common
+    # Dependency : ZenLib
+    # TODO: call the function who build a clean ZenLib compil archive
+    cp -r $ZL_source .
+
+    # Other Dependencies
+    mkdir -p Shared/Project/
     cp -r $zlib_source Shared/Project
+    # TODO
     #cp -r $curl_source Shared/Project
+    # TODO: _Common files, currently an empty dir on the online archive
+    mkdir Shared/Project/_Common
 
     echo "2: remove what isn't wanted..."
     cd MediaInfoLib
-        rm -f .cvsignore .gitignore
-        rm -fr .git
-        rm -f ToDo.txt
-        rm -f ReadMe.txt
+        rm -fr .cvsignore .git*
+        rm -f ToDo.txt ReadMe.txt
+        rm -fr Release
+        #cd Release
+        #    rm -f CleanUp.bat Example.ogg ReadMe_DLL_Windows.txt
+        #    rm -f Release_DLL_GNU_Prepare.bat Release_Lib_GNU_Prepare.bat
+        #    rm -f Release_DLL_Windows_i386.bat Release_DLL_Windows_x64.bat
+        #    rm -f Release_Source.bat UpgradeVersion.sh
+        #cd ..
         rm -fr debian
         cd Project
             rm -f GNU/libmediainfo.dsc GNU/libmediainfo.spec
@@ -72,13 +83,6 @@ function _linux_compil () {
             rm -fr PureBasic
         cd ..
         rm -fr Contrib
-        rm -fr Release
-        #cd Release
-        #    rm -f CleanUp.bat Example.ogg ReadMe_DLL_Windows.txt
-        #    rm -f Release_DLL_GNU_Prepare.bat Release_Lib_GNU_Prepare.bat
-        #    rm -f Release_DLL_Windows_i386.bat Release_DLL_Windows_x64.bat
-        #    rm -f Release_Source.bat UpgradeVersion.sh
-        #cd ..
         cd Source
             rm -f Doc/setlocale.txt
             rm -fr Install
@@ -101,12 +105,154 @@ function _linux_compil () {
     if $MakeArchives; then
         echo "3: compressing..."
         cd $Path
-        mkdir archives
-        # To specify the compression level
-        #$(/bin/tar -cf MediaInfo_Lib_Source.tar MediaInfoLib/)
-        (GZIP=-9 tar -czf archives/MediaInfo_DLL_${Version}_GNU_FromSource.tgz MediaInfo_DLL_${Version}_GNU_FromSource)
-        (BZIP=-9 tar -cjf archives/MediaInfo_DLL_${Version}_GNU_FromSource.tbz MediaInfo_DLL_${Version}_GNU_FromSource)
+        if ! b.path.dir? archives; then
+            mkdir archives
+        fi
+        #(GZIP=-9 tar -czf archives/MediaInfo_DLL_${Version}_GNU_FromSource.tgz MediaInfo_DLL_${Version}_GNU_FromSource)
+        #(BZIP=-9 tar -cjf archives/MediaInfo_DLL_${Version}_GNU_FromSource.tbz MediaInfo_DLL_${Version}_GNU_FromSource)
         (XZ_OPT=-9e tar -cJf archives/MediaInfo_DLL_${Version}_GNU_FromSource.txz MediaInfo_DLL_${Version}_GNU_FromSource)
+    fi
+
+}
+
+function _windows_compil () {
+
+    echo
+    echo "Prepare the archive for compilation on Windows:"
+    echo "1: copy what is wanted..."
+
+    cd $Path
+    mkdir libmediainfo_${Version}_AllInclusive
+    cd libmediainfo_${Version}_AllInclusive
+
+    cp -r $MIL_source .
+
+    # Dependency : ZenLib
+    # TODO: call the function who build a clean ZenLib compil archive
+    cp -r $ZL_source .
+
+    # Dependency : zlib
+    cp -r $zlib_source .
+
+    echo "2: remove what isn't wanted..."
+    cd MediaInfoLib
+        rm -f .cvsignore .gitignore
+        rm -fr .git
+        rm -fr Release
+        rm -fr debian
+        cd Project
+            rm -fr Coverity PureBasic
+            rm -f Java/ReadMe.txt NetBeans/ReadMe.txt
+            rm -f MSCS2008/asp_net_web_application/Default.aspx
+            rm -f MSCS2008/asp_net_web_application/Web.config
+            rm -f MSCS2010/asp_net_web_application/Default.aspx
+            rm -f MSCS2010/asp_net_web_application/Web.config
+            rm -f MSVB2010/Example/My\ Project/Application.myapp
+            rm -f MSVB2010/Example/My\ Project/Settings.settings
+            rm -f MSVC2008/ShellExtension/dlldata.c
+            rm -f MSVC2008/ShellExtension/MediaInfo_InfoTip_Register.bat
+            rm -f MSVC2008/ShellExtension/MediaInfo_InfoTip_UnRegister.bat
+            rm -f MSVC2008/ShellExtension/MediaInfoShellExt_.rgs
+            rm -f MSVC2008/ShellExtension/MediaInfoShellExt.rgs
+            rm -f MSVC2010/ShellExtension/MediaInfo_InfoTip_Register.bat
+            rm -f MSVC2010/ShellExtension/MediaInfo_InfoTip_UnRegister.bat
+            rm -f MSVC2010/ShellExtension/MediaInfoShellExt_.rgs
+            rm -f MSVC2010/ShellExtension/MediaInfoShellExt.rgs
+            rm -f MSVC2012/ShellExtension/dlldata.c
+            rm -f MSVC2012/ShellExtension/MediaInfo_InfoTip_Register.bat
+            rm -f MSVC2012/ShellExtension/MediaInfo_InfoTip_UnRegister.bat
+            rm -f MSVC2012/ShellExtension/MediaInfoShellExt_.rgs
+            rm -f MSVC2012/ShellExtension/MediaInfoShellExt.rgs
+            rm -f MSVC2013/ShellExtension/dlldata.c
+            rm -f MSVC2013/ShellExtension/MediaInfo_InfoTip_Register.bat
+            rm -f MSVC2013/ShellExtension/MediaInfo_InfoTip_UnRegister.bat
+            rm -f MSVC2013/ShellExtension/MediaInfoShellExt_.rgs
+            rm -f MSVC2013/ShellExtension/MediaInfoShellExt.rgs
+        cd ..
+        cd Source
+            rm -f ThirdParty/aes-gladman/aes_amd64.asm
+            rm -f ThirdParty/aes-gladman/aes.txt
+            rm -f ThirdParty/aes-gladman/aes_x86_v1.asm
+            rm -f ThirdParty/aes-gladman/aes_x86_v2.asm
+            rm -f ThirdParty/aes-gladman/via_ace.txt
+        cd ..
+    cd ..
+
+    if $MakeArchives; then
+        echo "3: compressing..."
+        cd $Path
+        if ! b.path.dir? archives; then
+            mkdir archives
+        fi
+        7z a -t7z -mx=9 -bd archives/libmediainfo_${Version}_AllInclusive.7z libmediainfo_${Version}_AllInclusive >/dev/null
+    fi
+
+}
+
+function _linux_packages () {
+
+    echo
+    echo "Prepare the archive for Linux packages creation:"
+    echo "1: copy what is wanted..."
+
+    cd $Path
+    mkdir tmp
+    cd tmp
+    cp -r $MIL_source .
+
+    echo "2: remove what isn't wanted..."
+    cd MediaInfoLib
+        rm -fr .cvsignore .git*
+        rm -f ToDo.txt ReadMe.txt
+        rm -fr Release
+        cd Project
+            rm -fr Coverity PureBasic
+            rm -f Java/ReadMe.txt NetBeans/ReadMe.txt
+            rm -f MSCS2008/asp_net_web_application/Default.aspx
+            rm -f MSCS2008/asp_net_web_application/Web.config
+            rm -f MSCS2010/asp_net_web_application/Default.aspx
+            rm -f MSCS2010/asp_net_web_application/Web.config
+            rm -f MSVB2010/Example/My\ Project/Application.myapp
+            rm -f MSVB2010/Example/My\ Project/Settings.settings
+            rm -f MSVC2008/ShellExtension/dlldata.c
+            rm -f MSVC2008/ShellExtension/MediaInfo_InfoTip_Register.bat
+            rm -f MSVC2008/ShellExtension/MediaInfo_InfoTip_UnRegister.bat
+            rm -f MSVC2008/ShellExtension/MediaInfoShellExt_.rgs
+            rm -f MSVC2008/ShellExtension/MediaInfoShellExt.rgs
+            rm -f MSVC2010/ShellExtension/MediaInfo_InfoTip_Register.bat
+            rm -f MSVC2010/ShellExtension/MediaInfo_InfoTip_UnRegister.bat
+            rm -f MSVC2010/ShellExtension/MediaInfoShellExt_.rgs
+            rm -f MSVC2010/ShellExtension/MediaInfoShellExt.rgs
+            rm -f MSVC2012/ShellExtension/dlldata.c
+            rm -f MSVC2012/ShellExtension/MediaInfo_InfoTip_Register.bat
+            rm -f MSVC2012/ShellExtension/MediaInfo_InfoTip_UnRegister.bat
+            rm -f MSVC2012/ShellExtension/MediaInfoShellExt_.rgs
+            rm -f MSVC2012/ShellExtension/MediaInfoShellExt.rgs
+            rm -f MSVC2013/ShellExtension/dlldata.c
+            rm -f MSVC2013/ShellExtension/MediaInfo_InfoTip_Register.bat
+            rm -f MSVC2013/ShellExtension/MediaInfo_InfoTip_UnRegister.bat
+            rm -f MSVC2013/ShellExtension/MediaInfoShellExt_.rgs
+            rm -f MSVC2013/ShellExtension/MediaInfoShellExt.rgs
+        cd ..
+        cd Source
+            rm -f ThirdParty/aes-gladman/aes_amd64.asm
+            rm -f ThirdParty/aes-gladman/aes.txt
+            rm -f ThirdParty/aes-gladman/aes_x86_v1.asm
+            rm -f ThirdParty/aes-gladman/aes_x86_v2.asm
+            rm -f ThirdParty/aes-gladman/via_ace.txt
+        cd ..
+    cd ..
+
+    if $MakeArchives; then
+        echo "3: compressing..."
+        cd $Path
+        if ! b.path.dir? archives; then
+            mkdir archives
+        fi
+        cd tmp
+        #(GZIP=-9 tar -czf ../archives/libmediainfo_${Version}.tgz MediaInfoLib)
+        #(BZIP=-9 tar -cjf ../archives/libmediainfo_${Version}.tbz MediaInfoLib)
+        (XZ_OPT=-9 tar -cJf ../archives/libmediainfo_${Version}.txz MediaInfoLib)
     fi
 
 }
@@ -132,60 +278,68 @@ function btask.PrepareSource.run () {
     if b.opt.has_flag? --all; then
         AllTarget=true
     fi
+    CleanUp=true
+    if b.opt.has_flag? --no-cleanup; then
+        CleanUp=false
+    fi
     MakeArchives=true
     if b.opt.has_flag? --no-archives; then
         MakeArchives=false
     fi
-    #CleanUp=true
-    #if [ $(b.opt.get_opt --no-cleanup) ]; then
-    #    CleanUp=false
-    #fi
 
-    #mkdir $Path
+    if ! b.path.dir? $Path || ! b.path.writable? $Path; then
+        b.abort "The repertory $Path doesn't exit or isn't writable."
+    fi
     cd $Path
+
+    # Clean up
     rm -fr archives
     rm -fr MediaInfo_DLL_${Version}_GNU_FromSource
+    rm -fr libmediainfo_${Version}_AllInclusive
+    rm -fr tmp
 
     if $LinuxCompil || $WindowsCompil || $LinuxPackages || $AllTarget; then
+        #echo -n
         _get_source
     else
         echo "Besides --project and --version, you must specify at least"
         echo "one of this options:"
         echo
         echo "--linux-compil|-lc"
-        echo "              Create the archive for compilation on Linux"
+        echo "              Prepare the archive for compilation on Linux"
         echo
         echo "--windows-compil|-wc"
-        echo "              Create the archive for compilation on Windows"
+        echo "              Prepare the archive for compilation on Windows"
         echo
         echo "--linux-packages|-lp|--linux-package"
-        echo "              Create the archive for Linux packages creation"
+        echo "              Prepare the archive for Linux packages creation"
         echo
         echo "--all|-a"
-        echo "              Create all the targets for this project"
+        echo "              Prepare all the targets for this project"
     fi
     if $LinuxCompil; then
         _linux_compil
     fi
     if $WindowsCompil; then
-        echo _windows_compil
+        _windows_compil
     fi
     if $LinuxPackages; then
-        echo _linux_packages
+        _linux_packages
     fi
     if $AllTarget; then
         _linux_compil
-        echo _windows_compil
-        echo _linux_packages
+        _windows_compil
+        _linux_packages
     fi
-
-    #..\..\Shared\Binary\Windows_i386\7-zip\7z a -r -t7z -mx9 libmediainfo__AllInclusive.7z MediaInfoLib\* ZenLib\* zlib\*
-    # unix (look the man):
-    #7z a -t7z -m0=lzma -mx=9 -mfb=64 -md=32m -ms=on archive.7z dir1
-
     
     # Clean up
-    # TODO: ??? existing in Release_Source.bat but not in Release_Source.sh
+    # TODO: option -nc
+    if $CleanUp; then
+        cd $Path
+        rm -fr MediaInfo_DLL_${Version}_GNU_FromSource
+        rm -fr libmediainfo_${Version}_AllInclusive
+        rm -fr tmp
+    fi
 
     #unset -v MIL_files index MIL_source
 }
